+++ ixmp/testing.py
@@ -179,6 +179,10 @@ def make_dantzig(mp, solve=False):
         If not :obj:`False`, then *solve* is interpreted as a path to a
         directory, and the model ``transport_ixmp.gms`` in the directory is run
         for the scenario.
+
+    See also
+    --------
+    .DantzigModel
     """
     # add custom units and region for timeseries data
     try:
@@ -188,55 +192,11 @@ def make_dantzig(mp, solve=False):
         pass
     mp.add_region('DantzigLand', 'country')
 
-    # initialize a new (empty) instance of an `ixmp.Scenario`
+    # Initialize a new Scenario, and use the DantzigModel class' initialize()
+    # method to populate it
     annot = "Dantzig's transportation problem for illustration and testing"
-    scen = Scenario(mp, version='new', annotation=annot, **models['dantzig'])
-
-    # define sets
-    scen.init_set('i')
-    scen.add_set('i', ['seattle', 'san-diego'])
-    scen.init_set('j')
-    scen.add_set('j', ['new-york', 'chicago', 'topeka'])
-
-    # capacity of plant i in cases
-    # add parameter elements one-by-one (string and value)
-    scen.init_par('a', idx_sets='i')
-    scen.add_par('a', 'seattle', 350, 'cases')
-    scen.add_par('a', 'san-diego', 600, 'cases')
-
-    # demand at market j in cases
-    # add parameter elements as dataframe (with index names)
-    scen.init_par('b', idx_sets='j')
-    b_data = pd.DataFrame([
-        ['new-york', 325, 'cases'],
-        ['chicago', 300, 'cases'],
-        ['topeka', 275, 'cases'],
-    ], columns=['j', 'value', 'unit'])
-    scen.add_par('b', b_data)
-
-    # distance in thousands of miles
-    # add parameter elements as dataframe (with index names)
-    scen.init_par('d', idx_sets=['i', 'j'])
-    d_data = pd.DataFrame([
-        ['seattle', 'new-york', 2.5, 'km'],
-        ['seattle', 'chicago', 1.7, 'km'],
-        ['seattle', 'topeka', 1.8, 'km'],
-        ['san-diego', 'new-york', 2.5, 'km'],
-        ['san-diego', 'chicago', 1.8, 'km'],
-        ['san-diego', 'topeka', 1.4, 'km'],
-    ], columns='i j value unit'.split())
-    scen.add_par('d', d_data)
-
-    # cost per case per 1000 miles
-    # initialize scalar with a value and a unit
-    scen.init_scalar('f', 90.0, 'USD_per_km')
-
-    # initialize the decision variables and equations
-    scen.init_var('x', idx_sets=['i', 'j'])
-    scen.init_var('z', None, None)
-    scen.init_equ('cost')
-    scen.init_equ('demand', idx_sets=['j'])
-    scen.init_equ('supply', idx_sets=['i'])
+    scen = Scenario(mp, **models['dantzig'], version='new', annotation=annot,
+                    scheme='dantzig', with_data=True)
 
     # commit the scenario
     scen.commit("Import Dantzig's transport problem for testing.")
@@ -246,8 +206,7 @@ def make_dantzig(mp, solve=False):
 
     if solve:
         # Solve the model using the GAMS code provided in the `tests` folder
-        scen.solve(model=str(Path(solve) / 'transport_ixmp'),
-                   case='transport_standard')
+        scen.solve(model='dantzig', case='transport_standard')
 
     # add timeseries data for testing `clone(keep_solution=False)`
     # and `remove_solution()`