+++ tests/test_core.py
@@ -40,9 +40,13 @@ def test_new_scen(test_mp):
     scen = ixmp.Scenario(test_mp, *can_args, version='new')
     assert scen.version == 0
 
-    # Create a Scenario with scheme='MESSAGE'
-    scen2 = ixmp.Scenario(test_mp, model='foo', scenario='bar',
-                          scheme='MESSAGE', version='new')
+    # A scenario with scheme='MESSAGE' can only be created with a subclass
+    class Scenario(ixmp.Scenario):
+        pass
+
+    scen2 = Scenario(test_mp, model='foo', scenario='bar',
+                     scheme='MESSAGE', version='new')
+
     # JDBCBackend complains unless these items are added
     scen2.add_set('technology', 't')
     scen2.add_set('year', '2000')
@@ -54,9 +58,7 @@ def test_new_scen(test_mp):
     with pytest.raises(RuntimeError):
         ixmp.Scenario(test_mp, model='foo', scenario='bar')
 
-    # …but loading with any subclass of ixmp.Scenario is fine
-    class Scenario(ixmp.Scenario):
-        pass
+    # …but loading with a subclass of ixmp.Scenario is fine
     Scenario(test_mp, model='foo', scenario='bar')
 
 
@@ -447,7 +449,7 @@ def test_log_level_raises(test_mp):
     pytest.raises(ValueError, test_mp.set_log_level, level='foo')
 
 
-def test_solve_callback(test_mp, test_data_path):
+def test_solve_callback(test_mp):
     """Test the callback argument to Scenario.solve().
 
     In real usage, callback() would compute some kind of convergence criterion.
@@ -460,8 +462,7 @@ def test_solve_callback(test_mp, test_data_path):
     scen = make_dantzig(test_mp)
 
     # Solve the scenario as configured
-    solve_args = dict(model=str(test_data_path / 'transport_ixmp'),
-                      case='transport_standard', gams_args=['LogOption=2'])
+    solve_args = dict(model='dantzig', gams_args=['LogOption=2'])
     scen.solve(**solve_args)
 
     # Store the expected value of the decision variable, x